LDE-714: moved cube copy functions
